Move slash package.
