+++ src/main/kotlin/tv/blademaker/killjoy/Launcher.kt
@@ -31,13 +31,12 @@ import tv.blademaker.killjoy.apis.stats.Website
 import tv.blademaker.killjoy.framework.CommandRegistry
 import tv.blademaker.killjoy.listeners.MainListener
 import tv.blademaker.killjoy.prometheus.Prometheus
-import tv.blademaker.killjoy.slash.handler.DefaultSlashCommandHandler
-import tv.blademaker.killjoy.slash.handler.SlashCommandHandler
+import tv.blademaker.slash.handler.DefaultSlashCommandHandler
+import tv.blademaker.slash.handler.SlashCommandHandler
 import tv.blademaker.killjoy.utils.CooldownManager
 import tv.blademaker.killjoy.utils.Loaders
 import tv.blademaker.killjoy.utils.SentryUtils
 import tv.blademaker.killjoy.utils.extensions.isInt
-import tv.blademaker.killjoy.valorant.AgentAbility
 import tv.blademaker.killjoy.valorant.ValorantAgent
 import tv.blademaker.killjoy.valorant.ValorantMap
 import tv.blademaker.killjoy.valorant.ValorantWeapon
