+++ src/main/kotlin/tv/blademaker/killjoy/commands/game/AgentSlashCommand.kt
@@ -18,8 +18,8 @@ package tv.blademaker.killjoy.commands.game
 import net.dv8tion.jda.api.EmbedBuilder
 import tv.blademaker.killjoy.Launcher
 import tv.blademaker.killjoy.framework.ColorExtra
-import tv.blademaker.killjoy.slash.AbstractSlashCommand
-import tv.blademaker.killjoy.slash.SlashCommandContext
+import tv.blademaker.slash.AbstractSlashCommand
+import tv.blademaker.slash.SlashCommandContext
 import tv.blademaker.killjoy.valorant.ValorantAgent
 
 @Suppress("unused")
