+++ src/main/kotlin/tv/blademaker/killjoy/commands/game/NewsSlashCommand.kt
@@ -16,8 +16,8 @@
 package tv.blademaker.killjoy.commands.game
 
 import tv.blademaker.killjoy.apis.news.NewsRetriever
-import tv.blademaker.killjoy.slash.AbstractSlashCommand
-import tv.blademaker.killjoy.slash.SlashCommandContext
+import tv.blademaker.slash.AbstractSlashCommand
+import tv.blademaker.slash.SlashCommandContext
 
 @Suppress("unused")
 class NewsSlashCommand : AbstractSlashCommand("news") {
