+++ src/main/kotlin/tv/blademaker/killjoy/commands/game/TopSlashCommand.kt
@@ -18,8 +18,8 @@ package tv.blademaker.killjoy.commands.game
 import net.dv8tion.jda.api.EmbedBuilder
 import tv.blademaker.killjoy.apis.riot.RiotAPI
 import tv.blademaker.killjoy.apis.riot.entities.Region
-import tv.blademaker.killjoy.slash.AbstractSlashCommand
-import tv.blademaker.killjoy.slash.SlashCommandContext
+import tv.blademaker.slash.AbstractSlashCommand
+import tv.blademaker.slash.SlashCommandContext
 
 @Suppress("unused", "DuplicatedCode")
 class TopSlashCommand : AbstractSlashCommand("top") {
