+++ src/main/kotlin/tv/blademaker/slash/SlashSubCommand.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and limitations under the License.
  ******************************************************************************/
 
-package tv.blademaker.killjoy.slash
+package tv.blademaker.slash
 
 import net.dv8tion.jda.api.entities.Command
